chore(components): rename and move stack
