+++ src/scenes/Dashboard/ApplicationFormEditor/scenes/EditStepForm/components/NewFieldDialog.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useEffect, useState } from "react";
-import { Stack } from "@/components/Stack";
+import { Stack } from "@/components/ui/stack";
 import { Button } from "@/components/ui/button";
 import { useForm } from "react-hook-form";
 import { trpc } from "@/services/trpc";
